fix(backend): store image file size in bytes instead of megabytes

+++ backend/app/models.py
@@ -181,7 +181,7 @@ class Image(Base):
     version: Mapped[int] = mapped_column(Integer, nullable=False, default=1, server_default="1")
     width: Mapped[int | None] = mapped_column(Integer, nullable=True)
     height: Mapped[int | None] = mapped_column(Integer, nullable=True)
-    file_size: Mapped[float | None] = mapped_column(Float, nullable=True)
+    file_size: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     created_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now()
     )
